# docs/openapi.yaml
openapi: 3.0.3
info:
    title: Intelligent After-Sales Service and Warranty API
    version: 1.0.0
    description: |
        Contract generated from the Laravel routes, controllers, Form Requests, policies,
        resources, enums, and exception renderer. Browser clients authenticate through a
        stateful Laravel Sanctum session. Obtain the CSRF cookie before state-changing calls.
servers:
    - url: http://127.0.0.1:8000
      description: Local development
...
security:
    - cookieAuth: []
paths:
...
    /api/tickets:
        get:
            tags: [Tickets]
            summary: List tickets
            operationId: listTickets
            x-required-permission: tickets.view
            parameters:
                - { $ref: '#/components/parameters/Search' }
                - { name: client_id, in: query, schema: { type: integer } }
                - { name: product_id, in: query, schema: { type: integer } }
                - { name: warranty_id, in: query, schema: { type: integer } }
                - { name: assigned_technician_id, in: query, schema: { type: integer } }
                - { name: created_by, in: query, schema: { type: integer } }
                - { name: priority, in: query, schema: { $ref: '#/components/schemas/TicketPriority' } }
                - { name: status, in: query, schema: { $ref: '#/components/schemas/TicketStatus' } }
                - { name: source, in: query, schema: { $ref: '#/components/schemas/TicketSource' } }
                - { name: warranty_eligible, in: query, schema: { type: boolean } }
                - { name: received_from, in: query, schema: { type: string, format: date } }
                - { name: received_to, in: query, schema: { type: string, format: date }, description: Must be on or after received_from. }
                - {
                      name: sort,
                      in: query,
                      schema: { type: string, enum: [ticket_number, priority, status, received_at, closed_at, created_at] },
                  }
                - { $ref: '#/components/parameters/Direction' }
                - { $ref: '#/components/parameters/PerPage100' }
            responses:
                '200':
                    {
                        description: Paginated tickets.,
                        content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedEnvelope' } } },
                    }
                '403': { $ref: '#/components/responses/Forbidden' }
                '422': { $ref: '#/components/responses/ValidationError' }
        post:
            tags: [Tickets]
            summary: Create a SAV ticket
            description: The service validates that related client, product, warranty, and invoice item form a consistent purchase.
            operationId: createTicket
            x-required-permission: tickets.create
            requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TicketCreate' } } } }
            responses:
                '201':
                    {
                        description: Created ticket.,
                        content: { application/json: { schema: { $ref: '#/components/schemas/TicketEnvelope' } } },
                    }
                '403': { $ref: '#/components/responses/Forbidden' }
                '422': { $ref: '#/components/responses/ValidationError' }
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: laravel_session
            description: Stateful Laravel Sanctum session. Fetch `/sanctum/csrf-cookie`; send cookies and `X-XSRF-TOKEN` for mutations.
    parameters:
        Search: { name: search, in: query, required: false, schema: { type: string, maxLength: 100 } }
        Direction: { name: direction, in: query, required: false, schema: { type: string, enum: [asc, desc] } }
        PerPage100: { name: per_page, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100 } }
...
    responses:
...
        Forbidden:
            description: The authenticated user is not authorized.
            content:
                { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { message: This action is unauthorized. } } }
...
        ValidationError:
            description: Input or business-rule validation failed.
            content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } }
...
    schemas:
        Error:
            type: object
            required: [message]
            properties:
                message: { type: string }
        ValidationError:
            type: object
            required: [message, errors]
            properties:
                message: { type: string, example: The given data was invalid. }
                errors:
                    type: object
                    additionalProperties:
                        type: array
                        items: { type: string }
...
        PaginatedEnvelope:
            type: object
            required: [data, links, meta]
            properties:
                data: { type: array, items: { type: object, additionalProperties: true } }
                links:
                    type: object
                    properties:
                        first: { type: string, nullable: true }
                        last: { type: string, nullable: true }
                        prev: { type: string, nullable: true }
                        next: { type: string, nullable: true }
                meta:
                    type: object
                    properties:
                        current_page: { type: integer }
                        from: { type: integer, nullable: true }
                        last_page: { type: integer }
                        path: { type: string }
                        per_page: { type: integer }
                        to: { type: integer, nullable: true }
                        total: { type: integer }
...
        TicketPriority: { type: string, enum: [low, normal, high, urgent] }
        TicketSource: { type: string, enum: [store, phone, email, web] }
        TicketStatus:
            type: string
            enum:
                [
                    opened,
                    received,
                    awaiting_diagnosis,
                    diagnosing,
                    awaiting_customer_approval,
                    awaiting_part,
                    repairing,
                    testing,
                    repaired,
                    ready_for_pickup,
                    delivered,
                    closed,
                    cancelled,
                ]
...
        TicketCreate:
            type: object
            additionalProperties: false
            required: [client_id, product_id, title, problem_description]
            properties:
                client_id: { type: integer, minimum: 1 }
                product_id: { type: integer, minimum: 1 }
                warranty_id: { type: integer, nullable: true, minimum: 1 }
                invoice_item_id: { type: integer, nullable: true, minimum: 1 }
                title: { type: string, minLength: 3, maxLength: 255 }
                problem_description: { type: string, minLength: 3, maxLength: 10000 }
                priority: { $ref: '#/components/schemas/TicketPriority' }
                source: { $ref: '#/components/schemas/TicketSource' }
...
        Ticket:
            type: object
            required:
                [id, uuid, ticket_number, client_id, product_id, title, problem_description, priority, status, source, warranty_eligible]
            properties:
                id: { type: integer }
                uuid: { type: string, format: uuid }
                ticket_number: { type: string }
                client_id: { type: integer }
                product_id: { type: integer }
                warranty_id: { type: integer, nullable: true }
                invoice_item_id: { type: integer, nullable: true }
                title: { type: string }
                problem_description: { type: string }
                priority: { $ref: '#/components/schemas/TicketPriority' }
                status: { $ref: '#/components/schemas/TicketStatus' }
                source: { $ref: '#/components/schemas/TicketSource' }
                warranty_eligible: { type: boolean }
                created_by: { type: integer }
                assigned_technician_id: { type: integer, nullable: true }
                received_at: { type: string, format: date-time, nullable: true }
                closed_at: { type: string, format: date-time, nullable: true }
                status_history_count: { type: integer }
                client: { type: object, nullable: true, additionalProperties: true }
                product: { type: object, nullable: true, additionalProperties: true }
                warranty: { type: object, nullable: true, additionalProperties: true }
                invoice_item: { type: object, nullable: true, additionalProperties: true }
                created_by_user: { type: object, nullable: true, additionalProperties: true }
                assigned_technician: { type: object, nullable: true, additionalProperties: true }
                status_history: { type: array, items: { type: object, additionalProperties: true } }
                ticket_history: { type: array, items: { type: object, additionalProperties: true } }
                attachments: { type: array, items: { $ref: '#/components/schemas/Attachment' } }
                created_at: { type: string, format: date-time, nullable: true }
                updated_at: { type: string, format: date-time, nullable: true }
        TicketEnvelope:
            type: object
            required: [data]
            properties: { data: { $ref: '#/components/schemas/Ticket' } }
...
        Attachment:
            type: object
            required: [uuid, original_filename, mime_type, size]
            properties:
                id: { type: integer }
                uuid: { type: string, format: uuid }
                original_filename: { type: string }
                mime_type: { type: string }
                size: { type: integer, description: Bytes. }
                is_previewable_image: { type: boolean }
                uploaded_by:
                    {
                        type: object,
                        nullable: true,
                        additionalProperties: true,
                        description: Uploader summary; numeric ID is omitted in the client portal.,
                    }
                preview_url: { type: string, nullable: true }
                download_url: { type: string, nullable: true }
                created_at: { type: string, format: date-time, nullable: true }
...